[SDK] Unique InfSession devices (#59575)
Ensure that the `InferenceSession`'s devices are unique.
In order to enable the equality tests, add a `__hash__` method on
`Device`, which hashes the label and id.

As a drive by, make `setDevices` in CompiledModel.h more predictable.
Previously this would set the default device to whatever was the last
device in the devices list.
Now, instead it will prefer a GPU if present.
Filed GEX-2093 to follow up, or at least audit this.

Historical note: this started out as trying to always add a `CPU()`
device but that causes `torch.compile` model runs to fail with
ILLEGAL_ADDRESS in a way that I can't reproduce locally.

MAX_PYTHON_ORIG_REV_ID: cfb72841070987642046333326be58d1c9da64bb